New approach to copy and modify the NanoEvents array
Three issues:
- Shallow copies were not directly possible without loading every field
being materialized.
- Inserting a virtual array made the array materialize right away if it
was the first one being inserted
- Using the array which is getting virtual array inserted as an argument
in the generator of the virtual array created a memory leak in awkward

To get around these until awkward offers a better way of handling it,
introduce a new class that keeps the array around as a dict and only
converts it to one big awkward array when needed.